PickFur: outU/outV per-operand double cast for fmadds (49.23->50.15)
Target casts u2 and weights.z to double individually before multiply,
matching the single-rounded fmadds shape; the inner u0*wx+u1*wy stays
single.

+++ src/chara_fur.cpp
@@ -1494,10 +1494,10 @@ int CChara::CModel::PickFur(
 						           kCharaFurDepthScaleBase / (weights.z + weights.x + weights.y));
 
 						const double outU = static_cast<double>(static_cast<float>(
-						    static_cast<double>(verts[2].m_u * weights.z) +
+						    static_cast<double>(verts[2].m_u) * static_cast<double>(weights.z) +
 						    static_cast<double>(verts[0].m_u * weights.x + verts[1].m_u * weights.y)));
 						const double outV = static_cast<double>(static_cast<float>(
-						    static_cast<double>(verts[2].m_v * weights.z) +
+						    static_cast<double>(verts[2].m_v) * static_cast<double>(weights.z) +
 						    static_cast<double>(verts[0].m_v * weights.x + verts[1].m_v * weights.y)));
 
 						if (outWorldPos != 0) {
